Use a SHA until the next gem version is released
sigstore-ruby has fallen out of step with the upstream Sigstore servers.
Patches have been applied, but a new release needs to be made before
this Action can rely on the gem itself. In the meantime, we can point to
the specific SHA to get the desired behavior.

The intention of this change is for it to be temporary.